+++ views/index.ejs
@@ -33,8 +33,8 @@
                       <button type="submit" class="btn btn-default">搜索</button>
                     </form>
                     <div class="nav navbar-nav navbar-right navbar-form">
-                            <button class="btn btn-success">注册</button>
-                            <button class="btn btn-primary">登录</button>
+                            <a href='/register' class="btn btn-success">注册</a>
+                            <a href='/login' class="btn btn-primary">登录</a>
                     </div>
                     <ul class="nav navbar-nav navbar-right">
                       <li class="dropdown">
@@ -54,6 +54,6 @@
         <div class="text-center text-muted">
             我的github首页: <a href="https://github.com/ningqizheng">https://github.com/ningqizheng</a>
         </div>
-    <p><%=test%></p>
+    <!-- <p><%=test%></p> -->
 </body>
 </html>
